Add reset function
You can now reset the level with the "k" key. This keeps us from having
to constantly refresh the page, and it will be useful when we switch
between levels.

A miscellaneous change: Bullet/environment collision is now more
accurate on the x axis. Y remains unchanged, though it still needs work.