Complete collision update, make positioning universally center-based.
A few quirks of the system as it stands:
1. When ducking, your position remains where it was in the y, and the value passed into the
collider changes to compensate your ducked stance. Whether this is better than changing
the position itself is unknown atm. Maybe bundle into component?
2. Tables sink below the ground when flipped. This is acceptable as it will be fixed when
physics are added, as the table's rotation won't be directly manipulated anymore.
3. When moving at the edges of the car, your bounding box may reach beyond your position.
This is because position adjusting is done AFTER updating the collider. This can be easily
fixed, but I'm waiting because of the train car update. Probably will encapsulate component
manipulation to facilitate this.
4. Bullets will be drawn the frame they collide. This was already an issue, but it will be fixed
soon hopefully.