Add manifolds
The current system does not work because the way the collision resolution works,
it requires actual collision data; in other words, LACK of a collision is useless to a
manifold resolution. Long story short, the train collider needs to be changed from a
normal bounding box to four border boxes to get proper collision information.